Added properties to MO5.Object and fixed connect(): - getProperty(key) - setProperty(key, value) - hasProperty(key)
When a property is set or changed using setProperty, an "propertyChange"
event is triggered on the MO5.Object.

This is useful for observing changes to the object, e.g. a Canvas
could use this to get notified when a CanvasObject changes its
property "clickable" to true and only then add the object to its
internal click tracking (not yet implemented - this is just
an example!).

MO5.Object.prototype.connect() was fixed; when using it, not
event2 was triggered on obj2, but instead event1...